requirements.txt
openai-whisper==20250625
PyAudio~=0.2
